% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone}
\format{
Predict the age of abalone from physical measurements
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Abalone}
}
\usage{
data(abalone)
}
\description{
Abalone
}
\keyword{datasets}
