% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{best_split,DecisionTreeClassifier-method}
\alias{best_split,DecisionTreeClassifier-method}
\title{Best Split function}
\usage{
\S4method{best_split}{DecisionTreeClassifier}(object, X, y, parms)
}
\arguments{
\item{object}{DecisionTree object}

\item{X}{is data}

\item{y}{is class values}

\item{parms}{parms in function}
}
\value{
A list with: best_idx name of the feature with the best split or Null if it not be found
best_thr: threshold found in the best split, or Null if it not be found
}
\description{
Function to get best split in Decision Tree.
Find the best split for node. "Beast" means that the mean
of impurity is the least possible.
To find the best division. Let's iterate through all the features.
All threshold / feature pairs will be computed in the numerical
features. In the features that are not numerical,
We get the best group of possible values
will be obtained based on an algorithm with the function
get_levels_categoric
}
