% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{best_split}
\alias{best_split}
\title{An S4 method to best split}
\usage{
best_split(object, ...)
}
\arguments{
\item{object}{DecisionTree object}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
An S4 method to best split
}
