% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{breast}
\alias{breast}
\title{Breast}
\format{
: Diagnostic Wisconsin Breast Cancer Database
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)}
}
\usage{
data(breast)
}
\description{
Breast
}
\keyword{datasets}
