% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{calculate_gini}
\alias{calculate_gini}
\title{Function calculate gini}
\usage{
calculate_gini(column_factor)
}
\arguments{
\item{column_factor}{class values}
}
\description{
Function to calculate gini index.
Formula is: 1 - n:num_classes sum probabilitie_class ^ 2
}
