% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{check_value}
\alias{check_value}
\title{Check value in leaf}
\usage{
check_value(value, threshold)
}
\arguments{
\item{value}{is the value in leaf node}

\item{threshold}{in leaf node}
}
\value{
TRUE if <= in numeric or \%in\% in factor
}
\description{
Function to check value in leaf from numeric until character
}
