% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{check_xy_interface}
\alias{check_xy_interface}
\title{Ceck interface x y}
\usage{
check_xy_interface(x, y)
}
\arguments{
\item{x}{data without class labels}

\item{y}{values class}
}
\description{
Check interface
}
