% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{cluster_labels.model_sslr_fitted}
\alias{cluster_labels.model_sslr_fitted}
\title{Cluster labels}
\usage{
\method{cluster_labels}{model_sslr_fitted}(object, type = "class", ...)
}
\arguments{
\item{object}{model_sslr_fitted model built}

\item{type}{of predict in principal model: class, raw}

\item{...}{other parameters to be passed}
}
\description{
Get labels of clusters
raw returns factor or numeric values
}
