% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{cluster_labels}
\alias{cluster_labels}
\title{Get labels of clusters}
\usage{
cluster_labels(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other parameters to be passed}
}
\description{
Cluster labels
}
