% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\docType{class}
\name{DecisionTreeClassifier-class}
\alias{DecisionTreeClassifier-class}
\title{Class DecisionTreeClassifier}
\description{
Class DecisionTreeClassifier
Slots: max_depth, n_classes_, n_features_,  tree_, classes,  min_samples_split,
min_samples_leaf
}
