% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{fit.model_sslr}
\alias{fit.model_sslr}
\title{Fit with formula and data}
\usage{
\method{fit}{model_sslr}(object, formula = NULL, data = NULL, ...)
}
\arguments{
\item{object}{is the model}

\item{formula}{is the formula}

\item{data}{is the total data train}

\item{...}{unused in this case}
}
\description{
Funtion to fit through the formula
}
