% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{fit_decision_tree,DecisionTreeClassifier-method}
\alias{fit_decision_tree,DecisionTreeClassifier-method}
\title{Fit decision tree}
\usage{
\S4method{fit_decision_tree}{DecisionTreeClassifier}(
  object,
  X,
  y,
  min_samples_split = 20,
  min_samples_leaf = ceiling(min_samples_split/3),
  w = 0.5
)
}
\arguments{
\item{object}{A DecisionTreeClassifier object}

\item{X}{A object that can be coerced as data.frame. Training instances}

\item{y}{A vector with the labels of the training instances. In this vector
the unlabeled instances are specified with the value \code{NA}.}

\item{min_samples_split}{the minimum number of observations to do split}

\item{min_samples_leaf}{the minimum number of any terminal leaf node}

\item{w}{weight parameter ranging from 0 to 1}
}
\description{
method in class DecisionTreeClassifier used to build a Decision Tree
}
