% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{fit_decision_tree}
\alias{fit_decision_tree}
\title{An S4 method to fit decision tree.}
\usage{
fit_decision_tree(object, ...)
}
\arguments{
\item{object}{DecisionTree object}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
An S4 method to fit decision tree.
}
