% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{fit_x_u.model_sslr}
\alias{fit_x_u.model_sslr}
\title{Fit with x , y (labeled data) and unlabeled data (x_U)}
\usage{
\method{fit_x_u}{model_sslr}(object, x = NULL, y = NULL, x_U = NULL, ...)
}
\arguments{
\item{object}{is the model}

\item{x}{is a data frame or matrix with train dataset without objective feature.
X only have labeled data}

\item{y}{is objective feature with labeled values}

\item{x_U}{train unlabeled data without objective feature}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Funtion to fit with x and y and x_U.
Function calcule y with NA values and append in y param
}
