% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{fit_xy.model_sslr}
\alias{fit_xy.model_sslr}
\title{Fit with x and y}
\usage{
\method{fit_xy}{model_sslr}(object, x = NULL, y = NULL, ...)
}
\arguments{
\item{object}{is the model}

\item{x}{is a data frame or matrix with train dataset without objective feature.
X have labeled and unlabeled data}

\item{y}{is objective feature with labeled values and NA values in unlabeled data}

\item{...}{unused in this case}
}
\description{
Funtion to fit with x and y
}
