% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{get_centers.model_sslr_fitted}
\alias{get_centers.model_sslr_fitted}
\title{Cluster labels}
\usage{
\method{get_centers}{model_sslr_fitted}(object, ...)
}
\arguments{
\item{object}{model_sslr_fitted model built}

\item{...}{other parameters to be passed}
}
\description{
Get labels of clusters
raw returns factor or numeric values
}
