% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{get_centers}
\alias{get_centers}
\title{Get centers model of clustering}
\usage{
get_centers(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other parameters to be passed}
}
\description{
Centers clustering
}
