% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{get_class_max_prob}
\alias{get_class_max_prob}
\title{Get most frequented}
\usage{
get_class_max_prob(trees, input)
}
\arguments{
\item{trees}{trees list}

\item{input}{is input to be predicted}
}
\description{
Get value most frequented in vector
Used in predictions. It calls a predict with type = "prob"
in Decision Tree
}
