% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{get_class_mean_prob}
\alias{get_class_mean_prob}
\title{Get mean probability over all trees as prob vector}
\usage{
get_class_mean_prob(trees, input)
}
\arguments{
\item{trees}{trees list}

\item{input}{is input to be predicted}
}
\description{
Get mean probability over all trees as prob vector.
It calls a predict with type = "prob"
in Decision Tree
}
