% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{get_function}
\alias{get_function}
\title{FUNCTION TO GET FUNCTION METHOD}
\usage{
get_function(met)
}
\arguments{
\item{met}{character}
}
\value{
method_train (function)
}
\description{
FUNCTION TO GET FUNCTION METHOD SPECIFIC
}
