% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{get_levels_categoric}
\alias{get_levels_categoric}
\title{Function to get gtoup from gini index}
\usage{
get_levels_categoric(column, Y)
}
\arguments{
\item{column}{is the column}

\item{Y}{values}
}
\description{
Function to get group from gini index.
Used in categorical variable
From: https://freakonometrics.hypotheses.org/20736
}
