% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{get_most_frequented}
\alias{get_most_frequented}
\title{Get most frequented}
\usage{
get_most_frequented(elements)
}
\arguments{
\item{elements}{vector with values}
}
\description{
Get value most frequented in vector
Used in predictions
}
