% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{get_value_mean}
\alias{get_value_mean}
\title{Get value mean}
\usage{
get_value_mean(trees, input)
}
\arguments{
\item{trees}{trees list}

\item{input}{is input to be predicted}
}
\description{
Get value most frequented in vector
Used in predictions. It calls a predict with type = "numeric"
in Decision Tree
}
