% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{get_x_y}
\alias{get_x_y}
\title{FUNCTION TO GET REAL X AND Y WITH FORMULA AND DATA}
\usage{
get_x_y(form, data)
}
\arguments{
\item{form}{formula}

\item{data}{data values, matrix, dataframe..}
}
\value{
x (matrix,dataframe...) and y(factor)
}
\description{
FUNCTION TO GET REAL X AND Y WITH FORMULA AND DATA
}
