% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{gini_or_variance}
\alias{gini_or_variance}
\title{Gini or Variance by column}
\usage{
gini_or_variance(X)
}
\arguments{
\item{X}{column to calculate variance or gini}
}
\description{
function used to calculate the gini coefficient or
variance according to the type of the column. This function is called
for the creation of the decision tree
}
