% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{gini_prob}
\alias{gini_prob}
\title{Function to compute Gini index}
\usage{
gini_prob(y, classe)
}
\arguments{
\item{y}{values}

\item{classe}{classes}
}
\description{
Function to compute Gini index
From: https://freakonometrics.hypotheses.org/20736
}
