% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{grow_tree,DecisionTreeClassifier-method}
\alias{grow_tree,DecisionTreeClassifier-method}
\title{Function grow tree}
\usage{
\S4method{grow_tree}{DecisionTreeClassifier}(object, X, y, parms, depth = 0)
}
\arguments{
\item{object}{DecisionTree instance}

\item{X}{data values}

\item{y}{classes}

\item{parms}{parameters for grow tree}

\item{depth}{depth in tree}
}
\description{
Function to grow tree in Decision Tree
}
