% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{grow_tree}
\alias{grow_tree}
\title{An S4 method to grow tree.}
\usage{
grow_tree(object, ...)
}
\arguments{
\item{object}{DecisionTree object}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
An S4 method to grow tree.
}
