% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COREG.R
\name{knn_regression}
\alias{knn_regression}
\title{knn_regression}
\usage{
knn_regression(k, x, y, p)
}
\arguments{
\item{k}{parameter in KNN model}

\item{x}{data}

\item{y}{vector labeled data}

\item{p}{distance order}
}
\description{
create model knn
}
