% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{newDecisionTree}
\alias{newDecisionTree}
\title{Function to create DecisionTree}
\usage{
newDecisionTree(max_depth)
}
\arguments{
\item{max_depth}{max depth in tree}
}
\description{
Function to create DecisionTree
}
