% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{Class Node for Decision Tree}
\description{
Class Node for Decision Tree
Slots: gini, num_samples, num_samples_per_class, predicted_class_value, feature_index
threshold, left, right, probabilities
}
