% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\docType{class}
\name{nullOrNumericOrCharacter-class}
\alias{nullOrNumericOrCharacter-class}
\title{An S4 class to represent a class with more types values: null, numeric or character}
\description{
An S4 class to represent a class with more types values: null, numeric or character
}
