% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{predict,DecisionTreeClassifier-method}
\alias{predict,DecisionTreeClassifier-method}
\title{Function to predict inputs in Decision Tree}
\usage{
\S4method{predict}{DecisionTreeClassifier}(object, inputs, type = "class")
}
\arguments{
\item{object}{The Decision Tree object}

\item{inputs}{data to be predicted}

\item{type}{Is param to define the type of predict.
It can be "class", to get class labels
Or "prob" to get probabilites for class in each input.
Default is "class"}
}
\description{
Function to predict inputs in Decision Tree
}
