% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{predict,RandomForestSemisupervised-method}
\alias{predict,RandomForestSemisupervised-method}
\title{Function to predict inputs in Decision Tree}
\usage{
\S4method{predict}{RandomForestSemisupervised}(
  object,
  inputs,
  type = "class",
  confident = "max_prob",
  allowParallel = TRUE
)
}
\arguments{
\item{object}{The Decision Tree object}

\item{inputs}{data to be predicted}

\item{type}{class raw}

\item{confident}{Is param to define the type of predict.
It can be "max_prob", to get class with sum of probability is the maximum
Or "vote" to get the most frequented class in all trees.
Default is "max_prob"}

\item{allowParallel}{Execute Random Forest in parallel if doParallel is loaded.}
}
\description{
Function to predict inputs in Decision Tree
}
