% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COREG.R
\name{predict.COREG}
\alias{predict.COREG}
\title{Predictions of the COREG method}
\usage{
\method{predict}{COREG}(object, x, type = "numeric", ...)
}
\arguments{
\item{object}{Self-training model built with the \code{\link{COREG}} function.}

\item{x}{A object that is data}

\item{type}{of predict in principal model (numeric)}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned (numeric).
}
\description{
Predicts the label of instances according to the \code{COREG} model.
}
\details{
For additional help see \code{\link{COREG}} examples.
}
