% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMLeastSquaresClassifier.R
\name{predict.EMLeastSquaresClassifierSSLR}
\alias{predict.EMLeastSquaresClassifierSSLR}
\title{Predict EMLeastSquaresClassifierSSLR}
\usage{
\method{predict}{EMLeastSquaresClassifierSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict EMLeastSquaresClassifierSSLR
}
