% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMNearestMeanClassifier.R
\name{predict.EMNearestMeanClassifierSSLR}
\alias{predict.EMNearestMeanClassifierSSLR}
\title{Predict EMNearestMeanClassifierSSLR}
\usage{
\method{predict}{EMNearestMeanClassifierSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict EMNearestMeanClassifierSSLR
}
