% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyRegularizedLogisticRegression.R
\name{predict.EntropyRegularizedLogisticRegressionSSLR}
\alias{predict.EntropyRegularizedLogisticRegressionSSLR}
\title{Predict EntropyRegularizedLogisticRegressionSSLR}
\usage{
\method{predict}{EntropyRegularizedLogisticRegressionSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict EntropyRegularizedLogisticRegressionSSLR
}
