% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaplacianSVM.R
\name{predict.LaplacianSVMSSLR}
\alias{predict.LaplacianSVMSSLR}
\title{Predict LaplacianSVMSSLR}
\usage{
\method{predict}{LaplacianSVMSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict LaplacianSVMSSLR
}
