% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTSVM.R
\name{predict.LinearTSVMSSLR}
\alias{predict.LinearTSVMSSLR}
\title{Predict LinearTSVMSSLR}
\usage{
\method{predict}{LinearTSVMSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict LinearTSVMSSLR
}
