% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCNearestMeanClassifier.R
\name{predict.MCNearestMeanClassifierSSLR}
\alias{predict.MCNearestMeanClassifierSSLR}
\title{Predict MCNearestMeanClassifierSSLR}
\usage{
\method{predict}{MCNearestMeanClassifierSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict MCNearestMeanClassifierSSLR
}
