% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{predict.model_sslr_fitted}
\alias{predict.model_sslr_fitted}
\title{Predictions of model_sslr_fitted class}
\usage{
\method{predict}{model_sslr_fitted}(object, x, type = NULL, ...)
}
\arguments{
\item{object}{model_sslr_fitted model built.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix
with the distances between the unseen instances and the selected training instances,
or a matrix of instances.}

\item{type}{of predict in principal model: class, raw, prob, vote, max_prob, numeric}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
tibble or vector.
}
\description{
Predicts from model. There are different types: class, prob, raw
class returns tibble with one column
prob returns tibble with probabilities class columns
raw returns factor or numeric values
}
