% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{predict.RandomForestSemisupervised_fitted}
\alias{predict.RandomForestSemisupervised_fitted}
\title{Predictions of the SSLRDecisionTree_fitted method}
\usage{
\method{predict}{RandomForestSemisupervised_fitted}(object, x, type = "class", confident = "max_prob", ...)
}
\arguments{
\item{object}{RandomForestSemisupervised_fitted.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix
with the distances between the unseen instances and the selected training instances,
or a matrix of instances.}

\item{type}{of predict in principal model}

\item{confident}{Is param to define the type of predict.
It can be "max_prob", to get class with sum of probability is the maximum
Or "vote" to get the most frequented class in all trees.
Default is "max_prob"}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances according to the RandomForestSemisupervised_fitted model.
}
