% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNNRCE.R
\name{predict.snnrceG}
\alias{predict.snnrceG}
\title{Predictions of the SNNRCE method}
\usage{
\method{predict}{snnrceG}(object, D, ...)
}
\arguments{
\item{object}{model instance}

\item{D}{distance matrix}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predicts the label of instances according to the \code{snnrceG} model.
}
