% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{predict.SSLRDecisionTree_fitted}
\alias{predict.SSLRDecisionTree_fitted}
\title{Predictions of the SSLRDecisionTree_fitted method}
\usage{
\method{predict}{SSLRDecisionTree_fitted}(object, x, type = "class", ...)
}
\arguments{
\item{object}{model SSLRDecisionTree_fitted.}

\item{x}{A object that can be coerced as matrix.
Depending on how was the model built, \code{x} is interpreted as a matrix
with the distances between the unseen instances and the selected training instances,
or a matrix of instances.}

\item{type}{of predict in principal model}

\item{...}{This parameter is included for compatibility reasons.}
}
\value{
Vector with the labels assigned.
}
\description{
Predicts the label of instances SSLRDecisionTree_fitted model.
}
