% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSVM.R
\name{predict.TSVMSSLR}
\alias{predict.TSVMSSLR}
\title{Predict TSVMSSLR}
\usage{
\method{predict}{TSVMSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict TSVMSSLR
}
