% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USMLeastSquaresClassifier.R
\name{predict.USMLeastSquaresClassifierSSLR}
\alias{predict.USMLeastSquaresClassifierSSLR}
\title{Predict USMLeastSquaresClassifierSSLR}
\usage{
\method{predict}{USMLeastSquaresClassifierSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict USMLeastSquaresClassifierSSLR
}
