% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WellSVM.R
\name{predict.WellSVMSSLR}
\alias{predict.WellSVMSSLR}
\title{Predict WellSVMSSLR}
\usage{
\method{predict}{WellSVMSSLR}(object, x, ...)
}
\arguments{
\item{object}{is the object}

\item{x}{is the dataset}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
Predict WellSVMSSLR
}
