% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{predict_inputs,DecisionTreeClassifier-method}
\alias{predict_inputs,DecisionTreeClassifier-method}
\title{Predict inputs Decision Tree}
\usage{
\S4method{predict_inputs}{DecisionTreeClassifier}(object, inputs, type = "class")
}
\arguments{
\item{object}{DecisionTree object}

\item{inputs}{inputs to be predicted}

\item{type}{type prediction, class or prob}
}
\description{
Function to predict one input in Decision Tree
}
