% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{predict_inputs}
\alias{predict_inputs}
\title{An S4 method to predict inputs.}
\usage{
predict_inputs(object, ...)
}
\arguments{
\item{object}{DecisionTree object}

\item{...}{This parameter is included for compatibility reasons.}
}
\description{
An S4 method to predict inputs.
}
