% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRFClassifier.R
\name{predictions.GRFClassifierSSLR}
\alias{predictions.GRFClassifierSSLR}
\title{predictions unlabeled data}
\usage{
\method{predictions}{GRFClassifierSSLR}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other parameters to be passed}
}
\description{
Predictions
}
