% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{predictions.model_sslr_fitted}
\alias{predictions.model_sslr_fitted}
\title{Predictions of unlabeled data}
\usage{
\method{predictions}{model_sslr_fitted}(object, type = "class", ...)
}
\arguments{
\item{object}{model_sslr_fitted model built}

\item{type}{of predict in principal model: class, raw}

\item{...}{other parameters to be passed}
}
\description{
Predictions of unlabeled data (transductive)
raw returns factor or numeric values
}
