% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRFClassifier.R
\name{predictions}
\alias{predictions}
\title{predictions unlabeled data}
\usage{
predictions(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other parameters to be passed}
}
\description{
Predictions
}
