% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{print.model_sslr}
\alias{print.model_sslr}
\title{Print model SSLR}
\usage{
\method{print}{model_sslr}(object)
}
\arguments{
\item{object}{model_sslr object to print}
}
\description{
Print model SSLR
}
