% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{train_generic}
\alias{train_generic}
\title{FUNCTION TO TRAIN GENERIC MODEL}
\usage{
train_generic(y, ...)
}
\arguments{
\item{y}{(optional) factor (classes)}

\item{...}{list parms trControl (method...)}
}
\value{
model trained
}
\description{
FUNCTION TO TRAIN GENERIC MODEL
}
